requirements-dev.txt
black>=20.8b1
eth-brownie>=1.19.0,<2.0.0
